Move server dependencies to optional dependencies
Move all the dependencies that are only needed for running the main (aka
server) process but not for the job scripts into an optional
dependencies group "server".
This dramatically reduces dependencies for the user code, making
installation faster and reducing potential version conflicts.

So now, user code which only needs the client API can simply depend on
"cluster_utils" as before, but for running the cluster_utils
applications (`grid_search`/`hp_optimization`), one needs to install
"cluster_utils[server]".

Imports of these applications are wrapped with
`OptionalDependencyImport("server")`, to give the user a useful error
message in case the necessary dependencies have not been installed.  For
this, `OptionalDependencyImport` has been moved to `base.utils`, so it
can be imported without triggering any server dependency imports.
