+++ docs/installation.rst
@@ -7,13 +7,20 @@ Requirements
 
 cluster_utils requires Python version >= 3.8.
 
+.. _basic_installation:
 
 Basic Installation
 ==================
 
+**Important:** cluster_utils by default only installs the dependencies that are needed
+for the job scripts.  This is to avoid unnecessary dependencies in your user code
+(assuming you are using a separate virtual environment for it).
+**To be able to run the cluster_utils applications, you need to install it with the
+optional "server" dependencies:**
+
 .. code-block:: bash
 
-   pip install git+https://github.com/martius-lab/cluster_utils.git
+   pip install "cluster_utils[server] @ git+https://github.com/martius-lab/cluster_utils.git"
 
 
 .. _optional_dependencies:
@@ -22,12 +29,12 @@ Optional Dependencies
 =====================
 
 Some features require additional dependencies that are not installed by default.  They
-can be installed by specifying the corresponding "optional dependencies group" like
-this:
+can be installed by specifying the corresponding "optional dependencies group".  Keep in
+mind that you should always also include the "server" group:
 
 .. code-block:: bash
 
-   pip install "cluster[EXTRA] @ git+https://github.com/martius-lab/cluster_utils.git"
+   pip install "cluster_utils[server,EXTRA] @ git+https://github.com/martius-lab/cluster_utils.git"
 
 where ``EXTRA`` should be replaced by one of the following identifiers (or multiple,
 separated by commas):
@@ -38,10 +45,15 @@ separated by commas):
 
    * - Identifier
      - Needed for
+   * - **server**
+     - Basic dependencies for running cluster_utils applications (see
+       :ref:`basic_installation`).
    * - **report**
      - For generating PDF reports (see :doc:`report`).
    * - **nevergrad**
      - To use the *nevergrad* optimizer in ``hp_optimization``.
+   * - **all**
+     - Alias for 'server,report,nevergrad'.
    * - **docs**
      - For building the documentation.
    * - **dev**
